simple_bot.py
import os
import logging
import sys
from telegram.ext import Application, CommandHandler, MessageHandler, filters, ConversationHandler

# Enable logging
logging.basicConfig(
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
)
logging.getLogger("httpx").setLevel(logging.WARNING)

logger = logging.getLogger(__name__)

# Get bot token from environment
TELEGRAM_BOT_TOKEN = os.environ.get("TELEGRAM_BOT_TOKEN", "your_bot_token_here")

# Command handlers
async def start(update, context):
    """Send a message when the command /start is issued."""
    await update.message.reply_text(
        "سلام! من ربات ساخت حساب Gmail و کلید API Gemini هستم.\n"
        "از دستورات زیر می‌توانید استفاده کنید:\n"
        "/help - راهنمای دستورات\n"
        "/create - ساخت حساب جدید\n"
        "/status - بررسی وضعیت درخواست‌ها"
    )

async def help_command(update, context):
    """Send a message when the command /help is issued."""
    await update.message.reply_text(
        "دستورات موجود:\n"
        "/start - شروع کار با ربات\n"
        "/create - شروع فرآیند ساخت حساب Gmail و کلید API\n"
        "/status - بررسی وضعیت درخواست‌های شما\n"
        "/help - نمایش این راهنما\n\n"
        "توجه: ممکن است برای ساخت موفق حساب به پروکسی/VPN نیاز داشته باشید."
    )

async def create_command(update, context):
    """Start the account creation process."""
    # This is a placeholder for the actual implementation
    await update.message.reply_text(
        "درخواست ساخت حساب Gmail و کلید API شما ثبت شد.\n"
        "این فرآیند ممکن است چند دقیقه طول بکشد."
    )
    
    # Log the creation request
    logger.info(f"User {update.effective_user.id} requested account creation")
    
    # In the future, this would trigger the actual account creation process
    # For now, just confirm receipt
    await update.message.reply_text(
        "⚠️ توجه: این عملکرد در حال حاضر در دست توسعه است."
    )

async def status_command(update, context):
    """Check the status of account creation requests."""
    # This is a placeholder for the actual implementation
    await update.message.reply_text(
        "در حال حاضر هیچ درخواستی در صف نیست.\n"
        "⚠️ این عملکرد در حال حاضر در دست توسعه است."
    )

def main():
    """Start the bot."""
    # Create the Application
    application = Application.builder().token(TELEGRAM_BOT_TOKEN).build()

    # Add command handlers
    application.add_handler(CommandHandler("start", start))
    application.add_handler(CommandHandler("help", help_command))
    application.add_handler(CommandHandler("create", create_command))
    application.add_handler(CommandHandler("status", status_command))

    # Start the Bot
    application.run_polling(allowed_updates=Application.parse_update())
    
    return "Bot started successfully."

def start_polling():
    """Start the bot in non-blocking mode."""
    try:
        return main()
    except Exception as e:
        logger.error(f"Error starting bot: {str(e)}")
        return f"Failed to start bot: {str(e)}"

if __name__ == "__main__":
    main()